+++ inc/Plugin/AcfToolsOptionsSubPage.php
@@ -22,7 +22,9 @@ class AcfToolsOptionsSubPage extends ToolsOptionsSubPage {
 				[
 					'page_title' => $config['title'],
 					'menu_title' => $config['short_title'],
-				] + $options
+					'menu_slug' => $config['slug'],
+				],
+				$options
 			)
 		);
 	}